Include ephemeral models when dropping orphaned relations (#259)
* Include ephemeral models when dropping orphaned relations

When checking for orphaned relations to drop, include ephemeral models so potential existing objects are dropped. The edge case is when a model was previously materialized as a table or view and then later changed to ephemeral, the old object never gets cleaned up and remains in the state it was last updated.

* Retrigger CI

---------

Authored-by: Eugene Kim